Don't destroy if db object is undefined

+++ src/test/e2e/stores/client-application-store.e2e.test.js
@@ -14,8 +14,10 @@ beforeEach(async () => {
     clientApplicationsStore = stores.clientApplicationsStore;
 });
 
-afterEach(async () => {
-    await db.destroy();
+afterAll(async () => {
+    if (db) {
+        await db.destroy();
+    }
 });
 
 test("Should be able to keep track of what we've announced", async () => {
