+++ src/test/e2e/stores/setting-store.e2e.test.js
@@ -12,7 +12,11 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-    await db.destroy();
+    afterAll(async () => {
+        if (db) {
+            await db.destroy();
+        }
+    });
 });
 
 test('should have api secret stored', async () => {
